services/content_kb/requirements.txt
yadisk>=3.4.0
google-generativeai>=0.8.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
Pillow>=10.0.0
